Add package.js
This file describe a NodeJS package, it also handles dependencies
that can be installed with a simple `npm install` and gives the
user a simple interface to define "scripts" that can be run with
npm.

For instance here I've added a start scripts that runs
`bin/bot.js` and the user can use `npm start` to start the project.